Sketch applicative section of MapReduce case study
This section describes an applicative MapReduce. We use a Natural
Transformation to convert between Applicative and Monad but on
reflection the Applicative MapReduce is more general and should probably
be preferred.

Suggestions to finish the case study:

- more discussion on monoids in data science (HyperLogLog etc.)
- discussion of benchmarks
- testing and benchmarking harness in code repository

To expand on the last point, we should have more supporting code in the
advanced-scala-code repo to make testing and benchmarking easier.